+++ Examples/HelloWorld/CounterView.swift
@@ -3,13 +3,13 @@ import Aptabase
 
 struct CounterView: View {
     @State var count: Int = 0
-    
+
     var body: some View {
         VStack {
             Text("Count = \(count)")
             Button(action: {
                 self.count += 1
-                Aptabase.shared.trackEvent("Increment", with: ["count": self.count])
+                Aptabase.shared.trackEvent("Increment", with: ["count": .integer(self.count)])
             }) {
                 Text("Increment")
             }.padding()
@@ -17,8 +17,6 @@ struct CounterView: View {
     }
 }
 
-struct CounterView_Previews: PreviewProvider {
-    static var previews: some View {
-        CounterView()
-    }
+#Preview {
+    CounterView()
 }