+++ Examples/HelloWorldMac/HelloWorldMacApp.swift
@@ -5,12 +5,11 @@ import Aptabase
 struct HelloWorldMacApp: App {
     init() {
         Aptabase.shared.initialize(
-            appKey: "A-DEV-0000000000", 
-            // optionally track events as release, avoiding the default environment variable
-            options: InitOptions(trackingMode: .asRelease) 
+            appKey: "A-DEV-0000000000",
+            with: InitOptions(trackingMode: .release)
         )
     }
-    
+
     var body: some Scene {
         WindowGroup {
             CounterView()