+++ Package.swift
@@ -1,29 +1,22 @@
-// swift-tools-version: 5.7
-// The swift-tools-version declares the minimum version of Swift required to build this package.
+// swift-tools-version: 6.0
 
 import PackageDescription
 
 let package = Package(
     name: "Aptabase",
     platforms: [
-      .iOS(.v13),
-      .macCatalyst(.v13),
-      .macOS(.v10_15),
-      .watchOS(.v6),
-      .tvOS(.v13),
+        .iOS(.v17),
+        .macOS(.v14),
+        .watchOS(.v10),
+        .tvOS(.v17),
+        .visionOS(.v1),
     ],
     products: [
-        // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
             name: "Aptabase",
             targets: ["Aptabase"]),
     ],
-    dependencies: [
-      .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.3.0")
-    ],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "Aptabase",
             resources: [